pyproject.toml
[tool.poetry]
name = "wireguard-mesh"
version = "0.2.0"
description = "A Wireguard L2 Mesh Network Automation Tool"
authors = ["Phillip Sitbon <phillip.sitbon@gmail.com>"]
readme = "readme.md"
repository = "https://github.com/sitbon/wireguard-mesh"
homepage = "https://github.com/sitbon/wireguard-mesh"
license = "AGPLv3"
keywords = ["wireguard", "mesh", "vpn", "networking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
]

packages = [
    { include = "mesh" },
]

[tool.poetry.scripts]
mesh = "mesh.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
wireguard-tools = "^0.4.4"
attrs = "^23.1.0"
fabric = "^3.2.2"
pyyaml = "^6.0.1"
pulumi = {version = "^3.96.1", optional = true}

[tool.poetry.extras]
pulumi = ["pulumi"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
